Create index.html
This is a temporary record that holds all of these elected officials and the sources and evidence against them. It's not exhaustive, and I intend to rewrite this (probably in Rails!) to be easier for people to submit persons, photos, sources, as well as review them. This is not intended to be a witch-hunt - but it IS intended to provide a reliable, well-sourced record to be used to hold these individuals accountable in upcoming elections.